+++ cmd/apps/substreams_tier1.go
@@ -26,7 +26,6 @@ import (
 	"github.com/streamingfast/cli"
 	"github.com/streamingfast/dauth"
 	discoveryservice "github.com/streamingfast/dgrpc/server/discovery-service"
-	"github.com/streamingfast/dsession"
 	_ "github.com/streamingfast/dsession/local"
 	firecore "github.com/streamingfast/firehose-core"
 	"github.com/streamingfast/firehose-core/launcher"
@@ -264,10 +263,9 @@ func RegisterSubstreamsTier1App[B firecore.Block](chain *firecore.Chain[B], root
 				NominalCapacity:    viper.GetFloat64("substreams-tier1-cpu-eviction-nominal-capacity"),
 			}
 
-			sessionPlugin := viper.GetString("common-session-plugin")
-			sessionPool, err := dsession.New(sessionPlugin, appLogger)
+			sessionPool, err := newCommonSessionPool(appLogger)
 			if err != nil {
-				return nil, fmt.Errorf("unable to create session pool: %w", err)
+				return nil, err
 			}
 
 			return app.NewTier1(appLogger,